requirements.txt
Flask==2.3.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.99.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
